Split registerExecutionRoutes by URL prefix (421 → 30 lines)
26 routes across five distinct namespaces were registered in a
single sequential setupLocalAuth-style function. Each namespace
now has its own file under ./execution-routes/, and the entry
file is a 30-line shell that calls each register* in order:

- shared.ts          userIdFrom helper (claims OR session fallback)
- execution.ts       /api/execution/{prepare, preview, approve,
                     dispatch} + /api/execution/tasks(/:id)
- human-bridge.ts    /api/execution/human-bridge/{list, claim,
                     complete, abandon, requeue}
- approval.ts        /api/approval/{sweep, decide, notifications,
                     notifications/:id/read}
- workflow.ts        Demi-style: /api/workflow/{inbox/inspect,
                     classify, draft, scheduling, meeting-follow-up}
- operational.ts     Newo-style: /api/operational/{memory(2x),
                     follow-up, follow-up/tick, orchestrate,
                     defer(2x)}

No behavior change — same routes, same auth guards, same payload
contracts. Comments added on the non-obvious bits (why /preview
has no auth, why admin notifications skip recipient_id, why the
prepare-then-watchdog ordering matters).
